srcs/billing-app/requirements.txt
pika==1.3.1
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.19
